- updated Makefile - applied norm to `expand_hashtable` - refactor ft_strlen: return type to size_t

+++ utils.c
@@ -25,9 +25,9 @@ t_entry	*new_entry(size_t key, char *value)
 	return (entry);
 }
 
-int	ft_strlen(char *str)
+size_t	ft_strlen(char *str)
 {
-	int	i;
+	size_t	i;
 
 	if (!str)
 		return (0);
